Make a proper FireCrown theory model. Compatible with FireCrown 1.4.0
